Weekly Summary
The objective is to create a student enrollment model for an education institute
Ask the user to specify the subfolder in the current working directory, where the files are stored. 2. Ask the user for a course number, outputting the number of eligible students in response. Incorrect course numbers should be ignored and the program should go on.  When the user does not specify any course number, pressing enter instead, the program should stop.